Modernize (#2)
Update to work with ZODB 5.

Drop support for ZooKeeper. It depended on an implementation that was too closely coupled to ZEO implementation that has changed.  It's doubtful that anyone was depending on this.

* drop version argument from invalidate, and update extension

It hasn't been required in a very long time and isn't allowed in ZODB5.

Update handling of extension data to deal transaction changes.

* Removed some uninteresting doctest output that varied accross ZODB4 and ZODB5.

* Fixed to work when called via multiprocessing.Process

Specifically in the case where the reactor was started in the parent process,
typically in a test.

* Drop ZooKeeper support

* moved changes to history

* ZODB4 still works

* added Travis tests

* updated package metadata

* adjust time zone to avoid spurious errors

* Fixed time setup not to depend on local time

And fixed a bare except